Dynamic Dropdown

+++ RARIndia.Model/Model/DataTableModel.cs
@@ -5,8 +5,8 @@
         public string SearchBy { get; set; }
         public string SortByColumn { get; set; }
         public string SortBy { get; set; }
-        public short PageIndex { get; set; } = 1;
-        public short PageSize { get; set; } = 10;
+        public int PageIndex { get; set; } = 1;
+        public int PageSize { get; set; } = 10;
         public string SelectedCentreCode { get; set; } = string.Empty;
         public int SelectedDepartmentID { get; set; }
     }
