+++ RARIndia.ViewModel/ViewModel/PageListViewModel.cs
@@ -3,7 +3,7 @@
     public class PageListViewModel
     {
         public int Page { get; set; }
-        public short RecordPerPage { get; set; }
+        public int RecordPerPage { get; set; }
         public int TotalResults { get; set; } = 0;
         public int TotalPages { get; set; } = 0;
         public int TotalRecordCount { get; set; } = 0;
